added an example using the tls_scheduler
