+++ include/beman/task/detail/promise_env.hpp
@@ -14,13 +14,13 @@ template <typename Promise>
 struct promise_env {
     const Promise* promise;
 
-    auto query(::beman::execution::get_allocator_t) const noexcept -> typename Promise::allocator_type {
+    auto query(::beman::execution::get_allocator_t const&) const noexcept -> typename Promise::allocator_type {
         return this->promise->get_allocator();
     }
-    auto query(::beman::execution::get_scheduler_t) const noexcept -> typename Promise::scheduler_type {
+    auto query(::beman::execution::get_scheduler_t const&) const noexcept -> typename Promise::scheduler_type {
         return this->promise->get_scheduler();
     }
-    auto query(::beman::execution::get_stop_token_t) const noexcept -> typename Promise::stop_token_type {
+    auto query(::beman::execution::get_stop_token_t const&) const noexcept -> typename Promise::stop_token_type {
         return this->promise->get_stop_token();
     }
